Revert conditional require of autoloader
If using Jentil as parent theme, Jentil is supposed to be installed
via `composer create-project`, in which case the autoloader would
exist in Jentil's root.

If installed as a package, it's supposed to be installed from
the consuming theme's root (eg: `wp-content/themes/my-theme),
and not from the web root.

In this case, Jentil's autoloader would not exist, but we
shouldn't get an error because Jentil's functions.php would not
be loaded.