Run Jentil in child theme if Jentil is installed as package
When Jentil is installed as theme, functions.php is loaded by
WordPress, so Jentil is run. As a package, we have to explicitly
run Jentil in our child theme.

We have added this to `run()` in Jentil's AbstractChildTheme
class to remove the necessity of doing so explicitly in every
child theme.