Remove style rules for shown elements in our toggle classes
The `display:initial;` which was used to show previously hidden
elements set them to `inline`, which is not what we may want for
all elements.

So now, we are only hiding elements when the appropriate class
is applied to those elements, and doing nothing when a class
responsible for showing the element is applied.

As long as the class used to hide the element has been removed,
the element would be shown in it's previous `display:<value>`.