Move composing classes up one level
This ensures they are in the same namespace as their peer classes
in that namespace, rather than being nested one level deeper into
their own namespace.

The advantage here is that, if ever we decide to add component
classes to a class, we do not need to change the namespace.

The previous structure required that we moved this new composing
class into it's own namespace.