Replace WP core's `jquery` with jQuery v3
The general advice is to avoid this, as it may lead to plugin
compatibility issues.

But, core's is outdated, with potential security issues.

I've taken care to implement this like WordPress did with the
outdated version (calling `jQuery.noConflict()` right after).

I've also enqueued the corresponding version of the jQuery Migrate
plugin.

I'm doing this on the front end alone, so we should not have any
issues at the back end (admin).

Everything seems to work OK, according to my tests. Will continue
to monitor for any issues.