Fix docker hub autobuild
Turns out "Dockerfile location" in autobuild config accepted
filename basename only.

So the trick was to set the "Build Context" to `/docker` where the
Dockerfiles reside, and use the basename of the Dockerfiles
(instead of context `/`, dockerfile `docker/<dockerfile>`)

While `/docker` is not the build context since our app's files
are one directory level above this directory, we override the
build context in our build hook, by setting that to `../`.

...and it works!!!

NOTE: The build context in the build hook seem to be relative to
the Dockerfile location.